lambda_train_docker/requirements_train.txt
scikit-learn==1.2.2
pandas==2.0.1
PyYAML==6.0